Fix desyncing forceMovements when loading new zones between them. Fix tutorial island hitpoints level xp gain reducing the player's level to 3 temporarily

+++ src/main/java/com/rs/game/model/entity/EntityHitBar.java
@@ -32,7 +32,7 @@ public class EntityHitBar extends HitBar {
 		int maxHp = entity.getMaxHitpoints();
 		if (hp > maxHp)
 			hp = maxHp;
-		return maxHp == 0 ? 0 : (hp * 255 / maxHp);
+		return maxHp == 0 ? 0 : (int) ((long)hp * 255 / maxHp);
 	}
 
 	@Override
